profile page, dropdown menu, some fixes
A new profile page with some account stats and a soon to be done delete button to delete your account. There is now a dropwdown menu to save some space on the right side of the navbar. Some fixes include alerting that the cart is empty instead of displaying an almost blank page when clicking on 'go to checkout' with an empty cart.